correct all hards
