--- author-PALAK AGRAWAL(1).c
@@ -1,17 +1,0 @@
-#include <stdio.h>
-int main()
-{
-	int arr={23,45,132,90,345,1,45,8};
-	int sum=sum_arr(arr[],8);
-	printf("Sum=%d",sum);	
-	return 0;
-}
-int sum_arr(int arr[],int n)
-{
-	int sum;
-	for(i=0;i<n;i++)
-	{
-		sum=sum+arr[i];
-	}
-	return sum;
-}
